[#24] Correct variable naming and add tostring for testcodes

+++ src/main/java/com/somaeja/post/mapper/PostMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface PostMapper {
 	int save(Post post);
 
-	Integer findPostById(Long postId);
+	Long findPostById(Long postId);
 
 	List<Post> findByAll();
